Prepare to break down action creator tests
